# Conflicts:
#	.travis.yml
#	setenv-android.sh

+++ cryptlib.cpp
@@ -148,7 +148,7 @@ size_t BlockTransformation::AdvancedProcessBlocks(const byte *inBlocks, const by
 	CRYPTOPP_ASSERT(outBlocks);
 	CRYPTOPP_ASSERT(length);
 
-	const ptrdiff_t blockSize = static_cast<ptrdiff_t>(BlockSize());
+	const size_t blockSize = BlockSize();
 	ptrdiff_t inIncrement = (flags & (BT_InBlockIsCounter|BT_DontIncrementInOutPointers)) ? 0 : blockSize;
 	ptrdiff_t xorIncrement = xorBlocks ? blockSize : 0;
 	ptrdiff_t outIncrement = (flags & BT_DontIncrementInOutPointers) ? 0 : blockSize;
@@ -331,7 +331,6 @@ void RandomNumberGenerator::GenerateIntoBufferedTransformation(BufferedTransform
 	}
 }
 
-/// \class ClassNullRNG
 /// \brief Random Number Generator that does not produce random numbers
 /// \details ClassNullRNG can be used for functions that require a RandomNumberGenerator
 ///   but don't actually use it. The class throws NotImplemented when a generation function is called.
